Update 2017_07_04_123910_create_quiz_runner_survey_user_table.php
Only add foreign keys if the referenced tables exist

+++ src/LaraQuiz/database/migrations/2017_07_04_123910_create_quiz_runner_survey_user_table.php
@@ -60,15 +60,19 @@ class CreateQuizRunnerSurveyUserTable extends Migration
 
             $table->primary(['survey_id', $this->userForeignKey]);
 
-            $table->foreign('survey_id')
-                ->references('id')->on($this->surveyTable)
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
+            if (Schema::hasTable($this->surveyTable)) {
+                $table->foreign('survey_id')
+                    ->references('id')->on($this->surveyTable)
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            }
 
-            $table->foreign($this->userForeignKey)
-                ->references('id')->on($this->userTable)
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
+            if (Schema::hasTable($this->userTable)) {
+                $table->foreign($this->userForeignKey)
+                    ->references('id')->on($this->userTable)
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            }
         });
     }
